+++ pkgs/cronet_http/tool/prepare_for_embedded.dart
@@ -29,6 +29,7 @@ import 'package:http/http.dart' as http;
 import 'package:xml/xml.dart';
 import 'package:yaml_edit/yaml_edit.dart';
 
+late final String _scriptName;
 late final Directory _packageDirectory;
 
 const _gmsDependencyName = 'com.google.android.gms:play-services-cronet';
@@ -42,16 +43,27 @@ final _cronetVersionUri = Uri.https(
   'dl.google.com',
   'android/maven2/org/chromium/net/group-index.xml',
 );
+// Finds the Google Play Services Cronet dependency line. For example:
+// '    implementation "com.google.android.gms:play-services-cronet:18.0.1"'
+final implementationRegExp = RegExp(
+  '^\\s*implementation [\'"]'
+  '$_gmsDependencyName'
+  ':\\d+.\\d+.\\d+[\'"]',
+  multiLine: true,
+);
 
 void main(List<String> args) async {
-  if (Directory.current.path.endsWith('tool')) {
-    _packageDirectory = Directory.current.parent;
-  } else {
-    _packageDirectory = Directory.current;
-  }
-
+  final script = Platform.script.toFilePath();
+  _scriptName = script.split(Platform.pathSeparator).last;
+  _packageDirectory = Directory(
+    Uri.directory(
+      '${script.replaceAll(_scriptName, '')}'
+      '..${Platform.pathSeparator}',
+    ).toFilePath(),
+  );
   final latestVersion = await _getLatestCronetVersion();
-  updateCronetDependency(latestVersion);
+  updateBuildGradle(latestVersion);
+  updateExampleBuildGradle();
   updatePubSpec();
   updateReadme();
   updateLibraryName();
@@ -75,22 +87,30 @@ Future<String> _getLatestCronetVersion() async {
 }
 
 /// Update android/build.gradle.
-void updateCronetDependency(String latestVersion) {
-  final fBuildGradle = File('${_packageDirectory.path}/android/build.gradle');
-  final gradleContent = fBuildGradle.readAsStringSync();
-  final implementationRegExp = RegExp(
-    '^\\s*implementation [\'"]'
-    '$_gmsDependencyName'
-    ':\\d+.\\d+.\\d+[\'"]',
-    multiLine: true,
-  );
+void updateBuildGradle(String latestVersion) {
+  final buildGradle = File('${_packageDirectory.path}/android/build.gradle');
+  final gradleContent = buildGradle.readAsStringSync();
   final newImplementation = '$_embeddedDependencyName:$latestVersion';
-  print('Patching $newImplementation');
+  print('Updating ${buildGradle.path}: adding $newImplementation');
   final newGradleContent = gradleContent.replaceAll(
     implementationRegExp,
     '    implementation "$newImplementation"',
   );
-  fBuildGradle.writeAsStringSync(newGradleContent);
+  buildGradle.writeAsStringSync(newGradleContent);
+}
+
+/// Remove the cronet reference from ./example/android/app/build.gradle.
+void updateExampleBuildGradle() {
+  final buildGradle =
+      File('${_packageDirectory.path}/example/android/app/build.gradle');
+  final gradleContent = buildGradle.readAsStringSync();
+
+  print('Updating ${buildGradle.path}: removing cronet reference');
+  final newGradleContent = gradleContent.replaceAll(
+    implementationRegExp,
+    '    // NOTE: removed in package:cronet_http_embedded',
+  );
+  buildGradle.writeAsStringSync(newGradleContent);
 }
 
 /// Update pubspec.yaml and example/pubspec.yaml.
@@ -120,7 +140,9 @@ void updateReadme() {
 void updateImports() {
   print('Updating imports in Dart files');
   for (final file in _packageDirectory.listSync(recursive: true)) {
-    if (file is File && file.path.endsWith('.dart')) {
+    if (file is File &&
+        file.path.endsWith('.dart') &&
+        !file.path.contains(_scriptName)) {
       final updatedSource = file.readAsStringSync().replaceAll(
             'package:cronet_http/cronet_http.dart',
             'package:cronet_http_embedded/cronet_http_embedded.dart',